requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-decouple
twilio
mercadopago